requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
html2text==2024.1.1
openai==1.59.0
python-dotenv==1.0.0
boto3==1.26.165
